// src/service/asset/IncidentApprove/list/type.ts
import { PageResponse } from '@/service/type'

export type ListIncidentApprove = {
  id: number
  code: string
  name: string
  reason: string
  severityManagement: {
    id: number
    name: string
    code: string
  }
  selfMaintenanceType: string
  incidentLocation: {
    id: number
    name: string
    code: string
  }
  recorder: {
    id: number
    code: string
    lastName: string
    firstName: string
    name: string
  }
  recordingTime: string
  actionType: string
  recordingStatus: string
  numberOfReviewType: string
  severityLevels: {
    severityManagement: {
      id: number
      name: string
      code: string
    }
    numberOfReviewType: string
  }[]

  isFinalReview: true
  approveStatus: string
  asset: {
    id: number
    name: string
    code: string
  }
  personApprove: {
    id: number
    code: string
    lastName: string
    firstName: string
    name: string
  }
  department: {
    id: number
    name: string
    code: string
  }
  note: string
}

export type RequestParams = {
  search?: string
  severityManagementId?: number
  departmentId?: number
  page: number
  size: number
}

export type ListForm = {
  search?: string
  severityManagement?: {
    id?: number
    name: string
    code: string
  }
  department?: {
    id?: number
    name: string
    code: string
  }
  page: number
  size: number
}

export type Response = {
  ListIncidentApprove: PageResponse<ListIncidentApprove[]>
}